+++ feature/login/sign-in-default/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 }
 
 android {
-    namespace = "com.gradation.lift.feature.login.termsOfUse"
+    namespace = "com.gradation.lift.feature.login.signInDefault"
     compileSdk = 33
 
     defaultConfig {
