+++ settings.gradle.kts
@@ -48,6 +48,12 @@ include(":feature:create-routine:common")
 
 include(":feature:login:sign-in")
 include(":feature:login:sign-up")
+include(":feature:login:sign-in-default")
+include(":feature:login:sign-up-default")
+include(":feature:login:terms-of-use")
+include(":feature:login:terms-of-use-detail")
+include(":feature:login:verify-email")
+include(":feature:login:reset-password")
 include(":feature:login:find-email-password")
 include(":feature:login:complete")
